Np 46642 nvi menu component (#5751)
* Reuse NVI candidates query hook

* Sopprt link back to correct nvi search

* Reorder hook params

* Update hook usage

* Remove wrongly committed functionality

* Update param name

* Move assignee filter to main content

* Move status filter to url param

* Fix styling of assignee selector

* Move all filter state to url params

* Move year filter

* FIx layout

* Simplify styling

* Ensure nvi filter always is valid

* Update type usages

* Minor cleanup

* Add data-testid

* Improve type

* Remove redundant export

* Add affiliation selector for area of responsibility

* Remove old area of responsibility selector

* Minor simplifications

* Optional sx prop

* Start splitting NVI menu into its own component

* Use nvi params hook

* Handle pagination reset (for most cases)

* Reset pagination for all filters

* Show correction list for NVI curator

* Ensure NVI status is showing for the correct year

* Set default value for status filter

* Add loading indicator for NVI status panel

* Use hook instead of prop

* Include forgotten file
